backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
alembic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
fpdf2
opencv-python-headless
numpy
